feat(chapter08): add distance calculation and track summary functions
