+++ chapter08/src/case01/README.md
@@ -67,5 +67,12 @@ class AccountType:
 
 ## 예시
 
+### 예시 (1)
+
+중첩함수는 숨겨진 데이터끼리 상호작용하기 쉬우므로 되도록 만들지 말자
+
+### 예시 (2)
+
+
 
 [^1]: 원래 함수를 소스 함수(_source function_)으로 일컫고, 복사해서 만든 함수를 타깃 함수(_target function_) 라고 한다
